+++ lib/ogr/feature_definition.rb
@@ -124,11 +124,7 @@ module OGR
     # @param index [Integer]
     # @return [OGR::GeometryFieldDefinition]
     def geometry_field_definition(index)
-      geometry_field_definition_ptr = FFI::OGR::API.OGR_FD_GetGeomFieldDefn(@c_pointer, index)
-
-      return nil if geometry_field_definition_ptr.null?
-
-      OGR::GeometryFieldDefinition.new(geometry_field_definition_ptr)
+      OGR::GeometryFieldDefinition.new_borrowed(FFI::OGR::API.OGR_FD_GetGeomFieldDefn(@c_pointer, index))
     end
 
     # @param name [String]